# admin-service/handlers/user_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// User represents a user data structure (placeholder)
// type User struct {
// 	ID   string `json:"id"`
// 	Name string `json:"name"`
// 	// Add other fields as needed
// }

// UserHandler handles user-related requests
// For now, it's a placeholder. In a real app, it would have a DB connection or service client.
type UserHandler struct {
	// db *gorm.DB // or a client to user-service
}

// NewUserHandler creates a new UserHandler
func NewUserHandler() *UserHandler {
	return &UserHandler{}
}

// GetUser godoc
// @Summary Get a user
// @Description Get details of a user by ID
// @Tags users
// @Accept  json
// @Produce  json
// @Param   id path string true "User ID"
// @Success 200 {object} map[string]interface{} "Successfully retrieved user"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "User not found"
// @Router /users/{id} [get]
func (h *UserHandler) GetUser(c *gin.Context) {
	id := c.Param("id")
	// In a real app, fetch user from user-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "GetUser called", "id": id})
}

// UpdateUser godoc
// @Summary Update a user
// @Description Update details of a user by ID
// @Tags users
// @Accept  json
// @Produce  json
// @Param   id path string true "User ID"
// @Param   user body map[string]interface{} true "User data to update"
// @Success 200 {object} map[string]interface{} "Successfully updated user"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "User not found"
// @Router /users/{id} [put]
func (h *UserHandler) UpdateUser(c *gin.Context) {
	id := c.Param("id")
	// var updatedUser User // Define your User struct if you have one
	// if err := c.ShouldBindJSON(&updatedUser); err != nil {
	// 	c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
	// 	return
	// }
	// In a real app, update user in user-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "UpdateUser called", "id": id /*, "updated_data": updatedUser*/})
}

// DeleteUser godoc
// @Summary Delete a user
// @Description Delete a user by ID
// @Tags users
// @Accept  json
// @Produce  json
// @Param   id path string true "User ID"
// @Success 200 {object} map[string]interface{} "Successfully deleted user"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "User not found"
// @Router /users/{id} [delete]
func (h *UserHandler) DeleteUser(c *gin.Context) {
	id := c.Param("id")
	// In a real app, delete user from user-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "DeleteUser called", "id": id})
}