# admin-service/handlers/video_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// Video represents a video data structure (placeholder)
// type Video struct {
// 	ID    string `json:"id"`
// 	Title string `json:"title"`
// 	// Add other fields as needed
// }

// VideoHandler handles video-related requests
// For now, it's a placeholder.
type VideoHandler struct {
	// db *gorm.DB // or a client to video-service
}

// NewVideoHandler creates a new VideoHandler
func NewVideoHandler() *VideoHandler {
	return &VideoHandler{}
}

// GetVideo godoc
// @Summary Get a video
// @Description Get details of a video by ID
// @Tags videos
// @Accept  json
// @Produce  json
// @Param   id path string true "Video ID"
// @Success 200 {object} map[string]interface{} "Successfully retrieved video"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "Video not found"
// @Router /videos/{id} [get]
func (h *VideoHandler) GetVideo(c *gin.Context) {
	id := c.Param("id")
	// In a real app, fetch video from video-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "GetVideo called", "id": id})
}

// UpdateVideo godoc
// @Summary Update a video
// @Description Update details of a video by ID
// @Tags videos
// @Accept  json
// @Produce  json
// @Param   id path string true "Video ID"
// @Param   video body map[string]interface{} true "Video data to update"
// @Success 200 {object} map[string]interface{} "Successfully updated video"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "Video not found"
// @Router /videos/{id} [put]
func (h *VideoHandler) UpdateVideo(c *gin.Context) {
	id := c.Param("id")
	// var updatedVideo Video // Define your Video struct
	// if err := c.ShouldBindJSON(&updatedVideo); err != nil {
	// 	c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
	// 	return
	// }
	// In a real app, update video in video-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "UpdateVideo called", "id": id /*, "updated_data": updatedVideo*/})
}

// DeleteVideo godoc
// @Summary Delete a video
// @Description Delete a video by ID
// @Tags videos
// @Accept  json
// @Produce  json
// @Param   id path string true "Video ID"
// @Success 200 {object} map[string]interface{} "Successfully deleted video"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "Video not found"
// @Router /videos/{id} [delete]
func (h *VideoHandler) DeleteVideo(c *gin.Context) {
	id := c.Param("id")
	// In a real app, delete video from video-service or DB
	c.JSON(http.StatusOK, gin.H{"message": "DeleteVideo called", "id": id})
}